requirements.txt
requests 
python-dotenv
tweepy
beautifulsoup4
datetime